Moving joins to their own test file
