+++ test/individualTests/individual.test.js
@@ -2,7 +2,6 @@ const assert = require('assert');
 const SQLParser = require('../../lib/SQLParser.js');
 const supportsArraySort = false;
 const {setup, disconnect, dbName} = require('../utils/mongo-client.js');
-const {queryResultTester} = require('../utils/query-tester');
 
 describe('Individual tests', function () {
     this.timeout(90000);
@@ -46,405 +45,6 @@ describe('Individual tests', function () {
         } catch (err) {
             return done(err);
         }
-    });
-
-    describe('joins', () => {
-        describe('left join', () => {
-            it('should be able to do a left join', async () => {
-                const queryText =
-                    'select * from orders as o left join `inventory` as i  on o.item=i.sku';
-                const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results);
-                    assert(results.length === 4);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-            it('should be able to do a left join with an unwind', async () => {
-                const queryText =
-                    'select * from orders as o left join `inventory|unwind` as i on o.item=i.sku';
-                const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results);
-                    assert(results.length === 4);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-
-            it('should be able to do a left join with an unwind in the query', async () => {
-                const queryText =
-                    'select o.id as OID,unwind(i) as inv from orders as o left join `inventory` i on o.item=i.sku';
-                /**
-                 * , unwind(`inventory`) as inventorySku
-                 */
-                const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results);
-                    assert(results.length === 4);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-
-            it('should be able to do a left join with an unwind and a case statement', async () => {
-                const queryText =
-                    "select (case when o.id=1 then 'Yes' else 'No' end) as IsOne from orders as o left join `inventory|unwind` as i  on o.item=i.sku";
-                const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-
-            it('should be able to do a left join on special characters', async () => {
-                const queryText =
-                    'select i.description,i.specialChars as iChars, o.item, o.specialChars as oChars from orders as o left join `inventory|unwind` as i on o.specialChars=i.specialChars';
-                const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results);
-                    assert(results.length === 4);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-
-            it('should be able to do a left join the other way round', async () => {
-                const queryText =
-                    'select * from orders as o left join `inventory` as i on i.sku=o.item';
-                const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results);
-                    assert(results.length === 4);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-            // it('should be able to do a left join with multiple conditions', async () => {
-            //     const queryText =
-            //         'select * from orders as o left join `inventory` as i on o.item=i.sku and o.id=i.id';
-            //     const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-            //     const hcp = [
-            //         {
-            //             $project: {
-            //                 o: '$$ROOT',
-            //             },
-            //         },
-            //         {
-            //             $lookup: {
-            //                 from: 'inventory',
-            //                 as: 'i',
-            //                 let: {
-            //                     o_item: '$o.item',
-            //                     o_id: '$o.id',
-            //                 },
-            //                 pipeline: [
-            //                     {
-            //                         $match: {
-            //                             $expr: {
-            //                                 $and: [
-            //                                     {
-            //                                         $eq: ['$sku', '$$o_item'],
-            //                                     },
-            //                                     {
-            //                                         $eq: ['$id', '$$o_id'],
-            //                                     },
-            //                                 ],
-            //                             },
-            //                         },
-            //                     },
-            //                 ],
-            //                 // pipeline: [
-            //                 //     {
-            //                 //         $match: {
-            //                 //             $expr: {
-            //                 //                 $and: [
-            //                 //                     {
-            //                 //                         $eq: ['$item', '$$i_sku'],
-            //                 //                     },
-            //                 //                     {
-            //                 //                         $eq: ['$id', '$$i_id'],
-            //                 //                     },
-            //                 //                 ],
-            //                 //             },
-            //                 //         },
-            //                 //     },
-            //                 // ],
-            //             },
-            //         },
-            //     ];
-            //     try {
-            //         const results = await mongoClient
-            //             .db(dbName)
-            //             .collection(parsedQuery.collections[0])
-            //             .aggregate(parsedQuery.pipeline)
-            //             .toArray();
-            //         assert(results);
-            //         assert(results.length === 4);
-            //         for (const result of results) {
-            //             assert(result.i.length);
-            //         }
-            //         return;
-            //     } catch (err) {
-            //         console.error(err);
-            //         throw err;
-            //     }
-            // });
-            // it('should be able to do a left join with multiple conditions the other way around', async () => {
-            //     const queryText =
-            //         'select * from orders as o left join `inventory` as i on i.sku=o.item and i.id=o.id';
-            //     const parsedQuery = SQLParser.makeMongoAggregate(queryText);
-            //     try {
-            //         const results = await mongoClient
-            //             .db(dbName)
-            //             .collection(parsedQuery.collections[0])
-            //             .aggregate(parsedQuery.pipeline)
-            //             .toArray();
-            //         assert(results);
-            //         assert(results.length === 4);
-            //         for (const result of results) {
-            //             assert(result.i.length);
-            //         }
-            //         return;
-            //     } catch (err) {
-            //         console.error(err);
-            //         throw err;
-            //     }
-            // });
-        });
-
-        describe('join order of queries', () => {
-            it('should prase the query that was not working on prod', async () => {
-                const queryText = `SELECT
-                unset(_id)
-                ,pol.CustId
-                ,cust.CustId as c_CustId
-                ,pol.ExecCode
-                ,emp.EmpCode as emp_EmpCode
-                from \`ams360-powerbi-basicpolinfo\` pol
-                INNER join \`ams360-powerbi-customer|unwind\` cust on pol.CustId = cust.CustId
-                LEFT join \`ams360-powerbi-employee|unwind\` emp on pol.ExecCode = emp.EmpCode
-                limit 5`;
-                const parsedQuery = SQLParser.parseSQL(queryText);
-                assert(parsedQuery);
-                const lookup1 = parsedQuery.pipeline[1].$lookup;
-                assert.deepEqual(lookup1, {
-                    from: 'ams360-powerbi-customer',
-                    foreignField: 'CustId',
-                    localField: 'pol.CustId',
-                    as: 'cust',
-                });
-                const lookup2 = parsedQuery.pipeline[4].$lookup;
-                assert.deepEqual(lookup2, {
-                    from: 'ams360-powerbi-employee',
-                    foreignField: 'EmpCode',
-                    localField: 'pol.ExecCode',
-                    as: 'emp',
-                });
-            });
-            describe('Existing example', () => {
-                it('should have the right lookup when no aliases specified', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on sku=item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-                it('should have the right lookup when only left alias is specified', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on i.sku=item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-                it('should have the right lookup when only right alias is specified', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on sku=o.item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-                it('should have the right lookup when both aliases are specified', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on i.sku=o.item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-            });
-            describe('Existing example - reversed', () => {
-                it('should have the right lookup when no aliases specified - reversed', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on item=sku';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'item',
-                        localField: 'o.sku',
-                        as: 'i',
-                    });
-                });
-                it('should have the right lookup when only left alias is specified - reversed', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on i.sku=item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-                it('should have the right lookup when only right alias is specified - reversed', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on sku=o.item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-                it('should have the right lookup when both aliases are specified - reversed', () => {
-                    const queryText =
-                        'select * from orders as o inner join `inventory` as i on i.sku=o.item';
-                    const parsedQuery = SQLParser.parseSQL(queryText);
-                    const lookup = parsedQuery.pipeline[1].$lookup;
-                    assert.deepEqual(lookup, {
-                        from: 'inventory',
-                        foreignField: 'sku',
-                        localField: 'o.item',
-                        as: 'i',
-                    });
-                });
-            });
-        });
-
-        describe('n-level joins', () => {
-            it('should be able to do n level joins without an unwind', async () => {
-                const queryText = `select
-                    o.id as orderId
-                    ,i.id as inventoryId
-                    ,c.id as customerId
-                    from orders as o
-                    inner join \`inventory\` as i
-                    on o.item=i.sku
-                    inner join \`customers\` as c
-                    on o.customerId=c.id
-                    where o.id=1`;
-                const parsedQuery = SQLParser.parseSQL(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results.length > 0);
-                    assert(results[0].orderId);
-                    assert(results[0].inventoryId);
-                    assert(results[0].customerId);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-            it('should be able to do n level joins with unwinds', async () => {
-                const queryText = `select
-                    o.id as orderId
-                    ,i.id as inventoryId
-                    ,c.id as customerId
-                    from orders as o
-                    inner join \`inventory|unwind\` as i
-                    on o.item=i.sku
-                    inner join \`customers|unwind\` as c
-                    on o.customerId=c.id`;
-                const parsedQuery = SQLParser.parseSQL(queryText);
-                try {
-                    const results = await mongoClient
-                        .db(dbName)
-                        .collection(parsedQuery.collections[0])
-                        .aggregate(parsedQuery.pipeline)
-                        .toArray();
-                    assert(results.length > 0);
-                    assert(results[0].orderId);
-                    assert(results[0].inventoryId);
-                    assert(results[0].customerId);
-                    return;
-                } catch (err) {
-                    console.error(err);
-                    throw err;
-                }
-            });
-        });
     });
 
     it('should be able to do a multipart-binary expression', async () => {
